Kernel: Cleaned everything
Clang format is not configured properly and makes a ton of code very
confusing. It removes indentations on switch statements, macros,
namespaces, and private/public. It also adds a line for if statements
when it could remain a one-liner. The inline assembly looks like a mess
and takes up ~4 lines for one thing. This makes things very hard to
understand what is going on. I also removed the CPUID functions as they
will not be used for a while and only print the output instead of
returning it.

Along with this commit I added a function to get the size of memory
which will be used during paging.

Signed-off-by: FiRe <levih@n11.dev>